+++ internal/gojs/compiler_test.go
@@ -65,12 +65,9 @@ func TestMain(m *testing.M) {
 		log.Println("gojs: skipping due missing Go binary:", err)
 		os.Exit(0)
 	}
-	start := time.Now()
 	if err = compileJsWasm(goBin); err != nil {
 		log.Panicln(err)
 	}
-	compilationTime := time.Since(start).Milliseconds()
-	log.Printf("gojs: compileFromGo took %dms and produced %dKB wasm", compilationTime, len(testBin)/1024)
 
 	// Define a compilation cache so that tests run faster. This works because
 	// all tests use the same binary.
@@ -83,7 +80,6 @@ func TestMain(m *testing.M) {
 
 	// Seed wazero's compilation cache to see any error up-front and to prevent
 	// one test from a cache-miss performance penalty.
-	start = time.Now()
 	rt := wazero.NewRuntime(testCtx)
 	defer rt.Close(testCtx)
 	_, err = rt.CompileModule(testCtx, testBin, wazero.NewCompileConfig())
@@ -91,8 +87,6 @@ func TestMain(m *testing.M) {
 		log.Panicln(err)
 	}
 	rt.Close(testCtx)
-	compilationTime = time.Since(start).Milliseconds()
-	log.Printf("gojs: compileFromWasm took %dms and produced %dKB cache", compilationTime, dirSize(compilationCacheDir)/1024)
 
 	// Configure fs test data
 	if d, err := fs.Sub(testFS, "sub"); err != nil {
@@ -101,20 +95,6 @@ func TestMain(m *testing.M) {
 		log.Panicln(err)
 	}
 	os.Exit(m.Run())
-}
-
-func dirSize(dir string) int64 {
-	var size int64
-	_ = filepath.Walk(dir, func(_ string, info os.FileInfo, err error) error {
-		if err != nil {
-			log.Panicln(err)
-		}
-		if !info.IsDir() {
-			size += info.Size()
-		}
-		return nil
-	})
-	return size
 }
 
 // compileJsWasm allows us to generate a binary with runtime.GOOS=js and
